+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ChangeLog history
 ## Features
 - write png files with Z corrdinate as file name
 - aligned block size option for generating tasks. This is useful to force alignment of chunks, and is used to align segmentation chunks for DVID.
+- add task number and current task index in global states. Thus, we know that a task could be the last one. This is useful to compute global statistics, such as maximum segmentation id in a segmentation and label indices of segmentation in DVID.
 
 ## Bug Fixes