+++ CHANGELOG.md
@@ -7,6 +7,8 @@ ChangeLog history
 ## Deprecations 
 
 ## Features
+- write png files with Z corrdinate as file name
+- aligned block size option for generating tasks. This is useful to force alignment of chunks, and is used to align segmentation chunks for DVID.
 - new operators to read and write NRRD file
 ## Bug Fixes